wip: monkey patch solution
