+++ sentry/hooks.py
@@ -4,11 +4,11 @@ from . import __version__ as app_version
 
 app_name = "sentry"
 app_title = "Sentry"
-app_publisher = "DigiThinkIT, Inc"
-app_description = "Logs errors to Sentry"
+app_publisher = "AgriTheory"
+app_description = "Error Reporting and Telemetry"
 app_icon = "octicon octicon-file-directory"
 app_color = "grey"
-app_email = "valmik@digithinkit.com"
+app_email = "support@agritheory.dev"
 app_license = "MIT"
 
 # Includes in <head>
