+++ setup.py
@@ -9,8 +9,8 @@ with open('requirements.txt') as f:
 _version_re = re.compile(r'__version__\s+=\s+(.*)')
 
 with open('sentry/__init__.py', 'rb') as f:
-    version = str(ast.literal_eval(_version_re.search(
-        f.read().decode('utf-8')).group(1)))
+	version = str(ast.literal_eval(_version_re.search(
+		f.read().decode('utf-8')).group(1)))
 
 setup(
 	name='sentry',
